fix bugs

+++ WpfApp1/WpfApp1/DeleteForm.xaml.cs
@@ -36,6 +36,27 @@ namespace WpfApp1
         {
             bool ok;
             string bookname = Bookname.Text.ToString().Trim();
+
+            string name;
+            string price;
+            string year;
+            string authorname;
+            string authorprofile;
+            string bookdescription;
+            bool isvip;
+            int salenumber;
+            int point;
+            string bookimagepath;
+            float vipfee;
+            string timefordiscount;
+            float discount;
+            int numberofpoints;
+            string pdfpath;
+            bool exist;
+
+            SQLmethodes.ReturnBookStats(0, bookname, out name, out authorname, out year, out price, out bookdescription, out authorprofile, out isvip, out salenumber, out point, out bookimagepath, out vipfee, out timefordiscount, out discount, out numberofpoints, out pdfpath, out exist);
+            if (!exist) return;
+
             SQLmethodes.DeleteBookFromBookTable(bookname, out ok);
             if (!ok) return;
             MessageBoxResult message = MessageBox.Show("Book Deleted Successfuly!");
